* initial update to 3.1.0 * code cleanup
git-svn-id: http://localhost:88/svn/uwow4/uwow2/trunk@417 3b753528-4e94-6941-993a-e4fe79dda6b7

+++ src/RealmServer/Objects/GameObjectFlags.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace Hazzik.Objects {
 	[Flags]
-	public enum GameObjectFlags : ushort {
+	public enum GameObjectFlags : uint {
 		InUse = 0x01,
 		Locked = 0x02,
 		ConditionalInteraction = 0x04,
